check age user added + update scene item user profile added + preview test files deleted
GeostoreFX: Check on the user's date of birth inserted to verify if the user is at least 13 years old: if it is affirmative, then it proceeds to the creation/modification of the user, otherwise it shows an error message. In addition, I inserted the management on the modification, on the modification of the object scene (item) user profile. Finally, deleted the test files