check duplicates + date updated + creating news for creating products/categories moved
duplicate checks for user, category and material objects added. Also, I improved the date get, so that the correct days and month can be output. Finally, I moved the news creation phase, so that it can be created only and exclusively if the products or categories have been created/deleted